chore: rename sample
