+++ sample/client/Cargo.toml
@@ -1,12 +1,15 @@
 [package]
-name = "server-sample"
+name = "client-sdk"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
 moqt = {path = "../../moqt"}
 tokio = { version = "1.45.0", features = ["full", "tracing"] }
-anyhow = { version = "1.0.75", features = ["backtrace"] }
 tracing = "0.1.37"
+async-trait = "0.1.74"
+anyhow = { version = "1.0.75", features = ["backtrace"] }
+uuid = { version = "1.18,1", features = ["v4"]}
+dashmap = "6.1.0"
 rcgen = "0.14.3"
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
