+++ sample/server/src/modules/core/publisher.rs
