+++ sample/server/src/modules/enums.rs
