+++ sample/server/src/modules/event_resolver.rs
