+++ sample/server/src/modules/event_resolver/sequence_handler.rs
