+++ sample/server/src/modules/manager.rs
