+++ sample/server/src/modules/repositories/subscriber_repository.rs
