--- sdk/client/Cargo.toml
@@ -1,4 +1,0 @@
-[package]
-name = "client-sdk"
-version = "0.1.0"
-edition = "2024"
