--- sdk/client/src/lib.rs
@@ -1,7 +1,0 @@
-pub struct ClientSDK {
-
-}
-
-impl ClientSDK {
-
-}
